requirements.txt
#  requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.6.0
autosarfactory @ git+https://github.com/girishchandranc/autosarfactory.git
lxml>=5.1.0
xmlschema>=3.3.0
typer>=0.12.0
rich>=13.7.0
pytest>=8.1.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
ruff>=0.4.0
httpx>=0.27.0
tenacity>=8.3.0

# NOTE: matlab.engine is NOT on PyPI — install manually:
#   cd "C:\Program Files\MATLAB\R2024b\extern\engines\python"
#   python setup.py install